cdk/requirements.txt
aws-cdk-lib>=2.84.0,<3.0.0  # Use a version compatible with schema 32.0.0
boto3
boto3-stubs
constructs>=10.0.0,<11.0.0
aws_cdk.aws_lambda_python_alpha
aws-cdk.aws-glue-alpha>=2.84.0a0
